pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payment-gateway"
version = "0.1.0"
description = "Reusable Django app for multiple payment providers (Razorpay, Stripe)"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Django>=3.2",
    "djangorestframework>=3.12",
    "stripe>=5.0",
    "razorpay>=1.0",
]

[project.urls]
"Homepage" = "https://example.com"

[tool.setuptools.packages.find]
where = ["."]
include = ["payment_gateway*"]
